+++ .gitignore
@@ -11,4 +11,6 @@ Thumbs.db
 *.log*
 node_modules/
 
-# Файлы сборки (допишите самостоятельно)
+# Файлы сборки
+public/bundle.js
+public/bundle.js.map
